Move logbook files into CaseDetail directory
