+++ src/app/features/cases/components/organisms/CaseDetail/utils.ts
@@ -26,3 +26,22 @@ export const getEigenaar = (caseData: Case) => {
 
   return undefined
 }
+
+export const isNullish = (a: any): a is undefined|null => a === undefined || a === null || a === ""
+
+const logbookTranslationsMap: Record<string, string> = {
+  nobody_present: "Niemand aanwezig",
+  no_cooperation: "Geen medewerking",
+  access_granted: "Toegang verleend",
+  malfunctioning_doorbell: "Bel functioneert niet",
+  intercom: "Contact via intercom",
+  hotel_furnished: "Hotelmatig ingericht",
+  vacant: "Leegstand",
+  likely_inhabited: "Vermoedelijk bewoond",
+  daytime: "Overdag",
+  weekend: "Weekend",
+  evening: "'s Avonds",
+  unknown: "Onbekend"
+}
+
+export const mapLogbookValue = (key: string): string => logbookTranslationsMap[key] ?? key
