--- src/app/features/cases/components/organisms/CaseLogbook/util.ts
@@ -1,18 +1,0 @@
-export const isNullish = (a: any): a is undefined|null => a === undefined || a === null || a === ""
-
-const translationMap: Record<string, string> = {
-  nobody_present: "Niemand aanwezig",
-  no_cooperation: "Geen medewerking",
-  access_granted: "Toegang verleend",
-  malfunctioning_doorbell: "Bel functioneert niet",
-  intercom: "Contact via intercom",
-  hotel_furnished: "Hotelmatig ingericht",
-  vacant: "Leegstand",
-  likely_inhabited: "Vermoedelijk bewoond",
-  daytime: "Overdag",
-  weekend: "Weekend",
-  evening: "'s Avonds",
-  unknown: "Onbekend"
-}
-
-export const translate = (key: string): string => translationMap[key] ?? key
